projekty/dmxperiments/requirements.txt
pydmx
pyserial
dmxpy